requirements.txt
annotated-types==0.6.0
anyio==3.7.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coverage==7.5.0
cryptography==41.0.4
Deprecated==1.2.14
fastapi==0.103.2
firebase-admin==6.2.0
gcloud==0.18.3
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.0
httpx==0.27.0
idna==3.4
iniconfig==2.0.0
jwcrypto==1.5.0
msgpack==1.0.7
oauth2client==4.1.3
packaging==24.0
pluggy==1.5.0
proto-plus==1.22.3
protobuf==4.25.0rc1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
pytest==8.1.2
pytest-asyncio==0.23.6
pytest-cov==5.0.0
python-dotenv==1.0.0
python-jwt==4.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==69.5.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
stripe==7.0.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
wrapt==1.15.0
